+++ IsekaiMod/Content/Features/IsekaiProtagonist/Villain/VillainProficiencies.cs
@@ -6,9 +6,9 @@ using Kingmaker.Blueprints.Items.Armors;
 using Kingmaker.Designers.Mechanics.Facts;
 using Kingmaker.UnitLogic.FactLogic;
 
-namespace IsekaiMod.Content.Features.IsekaiProtagonist.GodEmperor
+namespace IsekaiMod.Content.Features.IsekaiProtagonist.Villain
 {
-    class GodEmporerProficiencies
+    class VillainProficiencies
     {
         public static void Add()
         {
@@ -22,9 +22,9 @@ namespace IsekaiMod.Content.Features.IsekaiProtagonist.GodEmperor
 
             // Features
             var ExoticWeaponProficiency = Resources.GetModBlueprint<BlueprintFeature>("ExoticWeaponProficiency");
-            var GodEmporerProficiencies = Helpers.CreateBlueprint<BlueprintFeature>("GodEmporerProficiencies", bp => {
-                bp.SetName("God Emperor Proficiences");
-                bp.SetDescription("The God Emperor is proficient with all simple, martial, and exotic weapons and with all armor (heavy, light, and medium) and shields (including tower shields). They can cast {g|Encyclopedia:Spell}spells{/g} from this class while wearing armor and shields (including tower shields) without incurring the normal {g|Encyclopedia:Spell_Fail_Chance}arcane spell failure chance{/g}, but they incur the normal arcane spell failure chance for arcane spells received from other classes.");
+            var VillainProficiencies = Helpers.CreateBlueprint<BlueprintFeature>("VillainProficiencies", bp => {
+                bp.SetName("Villain Proficiences");
+                bp.SetDescription("The Villain is proficient with all simple, martial, and exotic weapons and with all armor (heavy, light, and medium) and shields (including tower shields). They can cast {g|Encyclopedia:Spell}spells{/g} from this class while wearing armor and shields (including tower shields) without incurring the normal {g|Encyclopedia:Spell_Fail_Chance}arcane spell failure chance{/g}, but they incur the normal arcane spell failure chance for arcane spells received from other classes.");
                 bp.m_Icon = null;
                 bp.AddComponent<AddFacts>(c => {
                     c.m_Facts = new BlueprintUnitFactReference[] {