+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.0] - 2022-11-05
+### Added
+- Added `Villain` archetype.
+- Added drake companion support for `ExpandedContent 0.4.40`.
+
 ## [1.3.0] - 2022-11-02
 ### Added
 - Added Mod support for `MysticalMayhem 0.1.3`.
@@ -34,7 +39,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.0.0] - 2022-10-28
 ### Added
 - Added `Isekai Protagonist` class
-- Added `God Emporer` archetype
+- Added `God Emperor` archetype
 - Added `Edge Lord` archetype
 - Added 3 Isekai backgrounds
 - Added related features and abilities